Refactor S3 flow log event retrieval to use retriever class
Replaced custom implementation of S3 flow log event processing with `NewS3FlowLogEventsRetriever`, consolidating logic for better maintainability and readability. Removed extensive, complex code handling S3 object processing, filtering, and streaming in favor of streamlined utility in `vpcflowlogs` package.
